Authentication
routes added and navigation fixed for redirection in case of logout and trying to access from url
