+++ frontend/src/components/pages/notfoundScreenF/notFound-page.jsx
@@ -1,0 +1,18 @@
+export default function NotFoundPage() {
+  return (
+    <>
+      <span
+        style={{ color: "#49243E" }}
+        className="lowercase font-welcomeheading text-5xl font-bold"
+      >
+        404
+      </span>
+      <span
+        style={{ color: "#49243E" }}
+        className="lowercase font-welcomeheading text-5xl font-bold"
+      >
+        page not found
+      </span>
+    </>
+  );
+}
