+++ frontend/src/components/pages/welcomeScreen/welcome-page.jsx
@@ -1,36 +1,46 @@
-import { SignInButton } from "@clerk/clerk-react";
+import { SignedOut, SignInButton, useUser } from "@clerk/clerk-react";
 import TypewriterString from "./typewritter";
 import { Button } from "@/components/ui/button";
 import Navbar from "../../navbar";
-import { Link } from "react-router-dom";
+import { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 
 export default function WelcomeScreen() {
+  
+  const { user } = useUser();
+  const navigate = useNavigate();
+  useEffect(() => {
+    if (user) {
+      navigate("/home");
+    }
+  }, [user, navigate]);
+
   return (
     <>
       <Navbar />
-      <span className="flex flex-col min-h-screen items-center justify-center">
-        <span
-          style={{ color: "#49243E" }}
-          className="lowercase font-welcomeheading text-5xl font-bold"
-        >
-          where&apos;s my{" "}
+      <SignedOut>
+        <span className="flex flex-col min-h-screen items-center justify-center">
           <span
-            style={{ textDecorationColor: "#BB8493" }}
-            className="underline"
+            style={{ color: "#49243E" }}
+            className="lowercase font-welcomeheading text-5xl font-bold"
           >
-            roommie
+            where&apos;s my{" "}
+            <span
+              style={{ textDecorationColor: "#BB8493" }}
+              className="underline"
+            >
+              roommie
+            </span>
+            ?
           </span>
-          ?
-        </span>
-        <TypewriterString />
-        <SignInButton>
-          <Link to="/home">
+          <TypewriterString />
+          <SignInButton mode="modal" afterSignInUrl="/home">
             <Button className="font-button text-lg">
               Let&apos;s get started!
             </Button>
-          </Link>
-        </SignInButton>
-      </span>
+          </SignInButton>
+        </span>
+      </SignedOut>
     </>
   );
 }
